+++ tests/Elements/HrTest.php
@@ -17,29 +17,29 @@ beforeEach(function() {
 test('test hr', function (): void {
     putenv('COLUMNS=20');
     $value = hr('Stay RAD!')->render();
-    $hr = "───StayRAD!────";
-    expect(str_replace([PHP_EOL, " "], "", strings($value)->trim()->toString()))->toEqual($hr);
+    $hr = "\e[0m───  Stay RAD!  ────" . PHP_EOL;
+    expect(strings($value)->toString())->toEqual($hr);
 });
 
 test('test hr without value', function (): void {
     putenv('COLUMNS=20');
     $value = hr()->render();
-    $hr = "────────────────────";
-    expect(str_replace([PHP_EOL, " "], "", strings($value)->trim()->toString()))->toEqual($hr);
+    $hr = "\e[0m────────────────────" . PHP_EOL;
+    expect(strings($value)->toString())->toEqual($hr);
 });
 
 test('test hr with text align left', function (): void {
     putenv('COLUMNS=20');
     $value = hr('Stay RAD!')->textAlignLeft()->render();
-    $hr = "───StayRAD!────";
-    expect(str_replace([PHP_EOL, " "], "", strings($value)->trim()->toString()))->toEqual($hr);
+    $hr = "\e[0m───  Stay RAD!  ────" . PHP_EOL;
+    expect(strings($value)->toString())->toEqual($hr);
 });
 
 test('test hr with text align right', function (): void {
     putenv('COLUMNS=20');
     $value = hr('Stay RAD!')->textAlignRight()->render();
-    $hr =  "────StayRAD!───";
-    expect(str_replace([PHP_EOL, " "], "", strings($value)->trim()->toString()))->toEqual($hr);
+    $hr =  "\e[0m────  Stay RAD!  ───" . PHP_EOL;
+    expect(strings($value)->toString())->toEqual($hr);
 });
 
 class HrTestTheme extends Theme implements ThemeInterface