+++ tests/Elements/ParagraphTest.php
@@ -8,5 +8,5 @@ use function Termage\paragraph;
 test('test paragraph', function (): void {
     putenv('COLUMNS=20');
     $value = paragraph('RAD')->render();
-    expect($value)->toBe("RAD                 " . PHP_EOL);
+    expect($value)->toBe("\e[0mRAD                 " . PHP_EOL);
 });